backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
python-jose[cryptography]==3.3.0
python-multipart>=0.0.12
groq>=1.1.0
python-dotenv>=1.0.1
pydantic>=2.9.0
pydantic-settings>=2.6.0
httpx>=0.27.0
bcrypt>=4.0.0
PyMuPDF>=1.24.0
python-docx>=1.1.0